test: organize tests in different files
